+++ src/main/java/rsystems/events/GuildChannelMoveEvent.java
@@ -6,7 +6,9 @@ import rsystems.SherlockBot;
 
 public class GuildChannelMoveEvent extends ListenerAdapter {
 
-
+/*
+Protection against moving a channel out of the archive category.
+ */
     public void onTextChannelUpdateParent( TextChannelUpdateParentEvent event) {
         if(event.getOldParent().getIdLong() == SherlockBot.guildMap.get(event.getGuild().getId()).getArchiveCategoryID()){
             SherlockBot.database.deleteRow("ArchiveTable","ChannelID",event.getChannel().getIdLong());