+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## 0.3.2
+Added failure handling to retrieve archive channel data.
+
+## 0.3.1
+Fixed archive channel permissions.  Sets overrides appropriately.
+
 ## 0.3.0
-Archive channels 
+Archive channels functionality
 
 ## 0.2.1
 Update the lookup command to provide more clear data.